requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
dj-database-url==0.4.2
Django==1.8
django-material==0.13.1
enum34==1.1.6
gunicorn==19.7.1
ipython==5.3.0
ipython-genutils==0.2.0
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-decouple==3.0
python-http-client==2.2.1
requests==2.13.0
scandir==1.5
sendgrid==3.6.5
sendgrid-django==4.0.4
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==3.3.0
